requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
gino==1.0.1
hiredis==1.1.0
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
